@@ -61,15 +61,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "9.1.11",
-    "@angular/common": "9.1.11",
-    "@angular/compiler": "9.1.11",
-    "@angular/core": "9.1.11",
-    "@angular/forms": "9.1.11",
-    "@angular/platform-browser": "9.1.11",
-    "@angular/platform-browser-dynamic": "9.1.11",
+    "@angular/animations": "10.0.0",
+    "@angular/common": "10.0.0",
+    "@angular/compiler": "10.0.0",
+    "@angular/core": "10.0.0",
+    "@angular/forms": "10.0.0",
+    "@angular/platform-browser": "10.0.0",
+    "@angular/platform-browser-dynamic": "10.0.0",
     "@angular/platform-server": "9.1.11",
-    "@angular/router": "9.1.11",
+    "@angular/router": "10.0.0",
     "@sentry/node": "^5.20.1",
     "@types/compression": "^1.7.0",
     "@types/cors": "^2.8.6",
@@ -93,19 +93,19 @@
     "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.901.9",
-    "@angular-devkit/build-ng-packagr": "~0.901.9",
-    "@angular/cli": "~9.1.9",
-    "@angular/compiler-cli": "~9.1.11",
-    "@angular/language-service": "~9.1.11",
-    "@nrwl/angular": "^9.4.4",
+    "@angular-devkit/build-angular": "0.1000.5",
+    "@angular-devkit/build-ng-packagr": "0.1000.5",
+    "@angular/cli": "10.0.5",
+    "@angular/compiler-cli": "10.0.0",
+    "@angular/language-service": "10.0.0",
+    "@nrwl/angular": "10.0.8",
     "@nrwl/builders": "^7.8.7",
-    "@nrwl/cypress": "9.4.4",
-    "@nrwl/eslint-plugin-nx": "9.4.4",
-    "@nrwl/express": "^9.5.0",
-    "@nrwl/jest": "9.4.4",
-    "@nrwl/node": "9.4.4",
-    "@nrwl/workspace": "9.4.4",
+    "@nrwl/cypress": "10.0.8",
+    "@nrwl/eslint-plugin-nx": "10.0.8",
+    "@nrwl/express": "10.0.8",
+    "@nrwl/jest": "10.0.8",
+    "@nrwl/node": "10.0.8",
+    "@nrwl/workspace": "10.0.8",
     "@types/chalk": "^2.2.0",
     "@types/express": "4.17.0",
     "@types/fs-extra": "^9.0.1",
@@ -160,7 +160,7 @@
     "tsickle": "^0.38.1",
     "tslib": "^1.13.0",
     "tslint": "~6.1.2",
-    "typescript": "~3.8.3",
+    "typescript": "3.9.7",
     "wait-on": "^5.0.1",
     "yamljs": "^0.3.0",
     "yargs": "^15.3.1"
