@@ -1,15 +1,11 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
-import {
-  ScullyDefaultSettings,
-  ScullyLibConfig,
-  SCULLY_LIB_CONFIG
-} from './config/scully-config';
+import { ScullyDefaultSettings, ScullyLibConfig, SCULLY_LIB_CONFIG } from './config/scully-config';
 import { IdleMonitorService } from './idleMonitor/idle-monitor.service';
 import { ScullyContentModule } from './scully-content/scully-content.module';
 
 @NgModule({
   imports: [ScullyContentModule],
-  exports: [ScullyContentModule]
+  exports: [ScullyContentModule],
 })
 export class ScullyLibModule {
   /**
@@ -19,12 +15,10 @@ export class ScullyLibModule {
    * there will be only 1 instance in our app.
    * We don't need forRoot, as we are not configuring anything in here.
    */
-  static forRoot(
-    config: ScullyLibConfig = ScullyDefaultSettings
-  ): ModuleWithProviders {
+  static forRoot(config: ScullyLibConfig = ScullyDefaultSettings): ModuleWithProviders<any> {
     return {
       ngModule: ScullyLibModule,
-      providers: [{ provide: SCULLY_LIB_CONFIG, useValue: config }]
+      providers: [{ provide: SCULLY_LIB_CONFIG, useValue: config }],
     };
   }
   constructor(private idle: IdleMonitorService) {}
